+++ packages/devui-vue/devui/tag-input/src/tag-input-types.ts
@@ -2,6 +2,7 @@ import type { ExtractPropTypes, PropType } from 'vue';
 
 export interface Suggestion {
   __index?: number;
+
   [x: string]: unknown;
 }
 
@@ -65,3 +66,10 @@ export const tagInputProps = {
 } as const;
 
 export type TagInputProps = ExtractPropTypes<typeof tagInputProps>;
+
+export type HandleEnter = () => void;
+export type OnSelectIndexChange = (isUp: boolean) => void;
+
+export interface UseInputKeydownReturnTypes {
+  onInputKeydown: (e: KeyboardEvent) => void;
+}
